refactor(meet): own participant connection lifecycle
