+++ frontend/src/apps/meet/utils/SFUMeetingManager.ts
@@ -3,7 +3,7 @@
  * Orchestrates SFU connection, media, and participant management
  *
  * This is a thin facade that coordinates the focused managers:
- * - SFUConnectionManager: Connection lifecycle and event handling
+ * - ParticipantConnection: Participant connection lifecycle and event handling
  * - SFUMediaManager: Producer/consumer operations
  * - SFURecoveryManager: ICE restart and recovery logic
  */
@@ -16,13 +16,12 @@ import type { ConnectionDetails, SFUClient } from "./SFUClient";
 import type { JoinRoomMediaState, JoinUserData } from "../types";
 import type { User } from "../composables/useCurrentUser";
 import {
-	SFUConnectionManager,
+	ParticipantConnection,
+	type ParticipantConnectionStartOptions,
+	type ParticipantConnectionState,
 	type SFUEventHandlers,
-} from "./sfu/SFUConnectionManager";
-import {
-	SFUMediaManager,
-	type PublishedMedia,
-} from "./sfu/SFUMediaManager";
+} from "./sfu/ParticipantConnection";
+import { SFUMediaManager, type PublishedMedia } from "./sfu/SFUMediaManager";
 import { SFURecoveryManager } from "./sfu/SFURecoveryManager";
 
 interface SFUMeetingManagerOptions {
@@ -39,7 +38,7 @@ export class SFUMeetingManager {
 	consumerManager: ConsumerManager;
 	transportManager: TransportManager;
 
-	connectionManager: SFUConnectionManager;
+	connectionManager: ParticipantConnection;
 	mediaManager: SFUMediaManager;
 	recoveryManager: SFURecoveryManager;
 
@@ -55,6 +54,8 @@ export class SFUMeetingManager {
 			sfuClient,
 			transportManager: this.transportManager,
 			meetingId: () => this.connectionManager?.meetingId ?? null,
+			schedule: (operation) =>
+				this.connectionManager.serializeTransportRecovery(operation),
 			onStarted: (reason) =>
 				this.connectionManager?.reportRecoveryState(
 					reason.includes("send") ? "recovering_send" : "recovering_receive",
@@ -65,21 +66,7 @@ export class SFUMeetingManager {
 			},
 			onFailed: async (_reason, result) => {
 				try {
-					const recoveries: Promise<unknown>[] = [];
-					if (result.send === "failed") {
-						this.connectionManager?.reportRecoveryState("recovering_send", _reason);
-						recoveries.push(this.mediaManager.rebuildSendSide());
-					}
-					if (result.recv === "failed") {
-						this.connectionManager?.reportRecoveryState("recovering_receive", _reason);
-						recoveries.push(this.connectionManager.resetReceiveSide());
-					}
-					const failedRecovery = (await Promise.allSettled(recoveries)).find(
-						(result) => result.status === "rejected",
-					);
-					if (failedRecovery?.status === "rejected") {
-						throw failedRecovery.reason;
-					}
+					await this.connectionManager.recoverFailedTransports(_reason, result);
 					this.connectionManager?.reportRecoveryState("healthy", _reason);
 				} catch (error) {
 					this.connectionManager?.reportRecoveryState("failed", _reason);
@@ -98,7 +85,7 @@ export class SFUMeetingManager {
 			() => this.connectionManager?.getCurrentUserId() ?? null,
 		);
 
-		this.connectionManager = new SFUConnectionManager({
+		this.connectionManager = new ParticipantConnection({
 			sfuClient,
 			videoManager: this.videoManager,
 			participantManager: this.participantManager,
@@ -121,6 +108,12 @@ export class SFUMeetingManager {
 		prefetchedDetails: ConnectionDetails | null = null,
 	): Promise<boolean> {
 		return this.connectionManager.connect(authToken, prefetchedDetails);
+	}
+
+	startParticipantConnection(
+		options: ParticipantConnectionStartOptions,
+	): Promise<ParticipantConnectionState> {
+		return this.connectionManager.start(options);
 	}
 
 	async joinRoom(
@@ -161,7 +154,7 @@ export class SFUMeetingManager {
 			mediaHandler.cleanup();
 			this.consumerManager.clear();
 			this.mediaManager.processedConsumers.clear();
-			this.connectionManager.bufferedProducerEvents = [];
+			this.connectionManager.clearBufferedReconciliationEvents();
 			this.transportManager.cleanup();
 
 			await this.transportManager.initializeDevice();
@@ -299,6 +292,10 @@ export class SFUMeetingManager {
 
 	get isConnected(): boolean {
 		return this.connectionManager.isConnected;
+	}
+
+	get participantConnectionState(): ParticipantConnectionState {
+		return this.connectionManager.state;
 	}
 
 	get eventTarget(): EventTarget {
